MDL-57698 backup: provide NO_LOCK hints on recordsets
When performing a SELECT...UPDATE loop on the same table, SQL Server
will deadlock (SELECT query in ASYNC_NETWORK_IO, UPDATE with an IX lock
on the #backup_ids_temp table).

Indicating in the SELECT that, despite being in the READ COMMITTED
isolation level, we don't mind reading uncommitted data (because we
*won't* -- we only go *forwards* through the recordset) avoids the
deadlock.